Implement methods to make unit tests pass.
NewChain() accepts a mode parameter. This describes the cardinality of
its chain keys, i.e. how many sequential tokens are used. Given the
input "The quick brown fox jumped over the lazy dog.", a modality of 1
would result in keys "the", "quick", "brown", and so on. A modality of 2
would produce keys "the,quick", "quick,brown", "brown,fox" and so on
into higher modalities.

Count() returns the total number of unique keys in the markov chain.
This isn't super important/functional but it's an easy attribute to use
for some crude unit tests.

Update() takes an input string and builds the probability tables. It
checks for existing keys, updates probability nodes or creates them as
appropriate.

Also added stubs for Speak() and Prompt(), the functions that will
generate output at random, or based on an input string respectively.